erase fwi_utils, add utils funciton to fwi and refactoring
